Update INV.R
The MakeCache matrix function creates a special matrix object that can cache its inverse.
It first checks to see if the inverse has already been calculated. If so, it gets the inverse from the cache and skips the computation. Otherwise, it calculates the inverse of the matrix and sets the value of the inverse matrix in the cache by using the setinverse function.
The operator <<-  is used to set the value to the variable from other environments.